chore: move to websocket package
