+++ src/test/kotlin/smalltalk/backend/apply/websocket/StompConnectionTests.kt
@@ -1,4 +1,4 @@
-package smalltalk.backend.apply
+package smalltalk.backend.apply.websocket
 
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.kotest.core.spec.style.FunSpec
